LPS-164607 Move to verify process
